+++ nuttx/configs/mbed/src/lpc17_leds.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * configs/mbed/src/up_leds.c
+ * configs/mbed/src/lpc17_leds.c
  *
  *   Copyright (C) 2010, 2013, 2015 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
@@ -57,7 +57,7 @@
 #ifdef CONFIG_ARCH_LEDS
 
 /****************************************************************************
- * Definitions
+ * Pre-processor Definitions
  ****************************************************************************/
 
 /* CONFIG_DEBUG_LEDS enables debug output from this file (needs CONFIG_DEBUG
